+++ internal/dao/server.go
@@ -92,6 +92,12 @@ func (s *Server) GetByFilter(_ context.Context, req *storagepb.DaoByFilterReques
 		})
 	}
 
+	if len(req.GetFungibleIds()) != 0 {
+		filters = append(filters, FungibleIDsFilter{
+			FungibleIDs: req.GetFungibleIds(),
+		})
+	}
+
 	list, err := s.sp.GetByFilters(filters)
 	if err != nil {
 		log.Error().Err(err).Msgf("get daos by filter: %+v", req)